pyproject.toml
[tool.pyright]
exclude = ["**/node_modules",
    "**/__pycache__",
    "src/pp_ml.py",
]
strict = ["src", "."]
venvPath = "."
venv = ".venv"
reportMissingTypeStubs = false
reportMissingImports = false

[tool.poetry]
name = "ia-cours-4"
version = "0.1.0"
description = ""
authors = ["Joé FEUCHT <71235356+joe67100@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
black = "^24.10.0"
pyright = "^1.1.389"
pre-commit = "^4.0.1"
picsellia = "^6.19.2"
ultralytics = "^8.3.40"
dynaconf = "^3.2.6"
pyyaml = "^6.0.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
